+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleStorageInterpreter.scala
@@ -1,6 +1,7 @@
 package org.broadinstitute.dsde.workbench
 package google2
 
+import java.nio.channels.Channels
 import java.nio.file.{Path, Paths}
 import java.time.Instant
 
@@ -13,7 +14,7 @@ import com.google.cloud.storage.BucketInfo.LifecycleRule
 import com.google.cloud.storage.Storage.{BlobListOption, BlobSourceOption, BlobTargetOption, BucketSourceOption}
 import com.google.cloud.storage.{Acl, Blob, BlobId, BlobInfo, BucketInfo, Storage, StorageOptions}
 import com.google.cloud.{Identity, Policy, Role}
-import fs2.{text, Stream}
+import fs2.{Stream, text}
 import io.chrisdavenport.log4cats.StructuredLogger
 import io.circe.Decoder
 import io.circe.fs2._
@@ -86,7 +87,18 @@ private[google2] class GoogleStorageInterpreter[F[_]: ContextShift: Timer: Async
         s"com.google.cloud.storage.Storage.get(${BlobId.of(bucketName.value, blobName.value)})"
       )
       r <- blobOpt match {
-        case Some(blob) => Stream.emits(blob.getContent()).covary[F]
+        case Some(blob) =>
+          // implementation based on fs-blobstore
+          fs2.io.readInputStream(
+            Channels.newInputStream {
+              val reader = blob.reader()
+              reader.setChunkSize(chunkSize)
+              reader
+            }.pure[F],
+            chunkSize,
+            blocker,
+            closeAfterUse = true
+          )
         case None       => Stream.empty
       }
     } yield r
@@ -425,6 +437,8 @@ private[google2] class GoogleStorageInterpreter[F[_]: ContextShift: Timer: Async
     case None    => blocker.blockOn(fa)
     case Some(s) => s.withPermit(blocker.blockOn(fa))
   }
+
+  private val chunkSize = 1024 * 1024 * 2 // com.google.cloud.storage.BlobReadChannel.DEFAULT_CHUNK_SIZE
 }
 
 object GoogleStorageInterpreter {